agentgpt/requirements.txt
openai
PyPDF2
pdfminer.six